Python Codes/requirements.txt
regex
azure-ai-textanalytics
azure-cognitiveservices-language-textanalytics==0.2.0
azure-ai-textanalytics --pre
time